WKActionSheet should dismiss with animation when done with the sheet https://bugs.webkit.org/show_bug.cgi?id=167804 <rdar://problem/30334861>
Reviewed by Tim Horton.

Currently, we do not dismiss the action menu when -doneWithSheet is invoked. While this is okay when the menu is
dismissed via tap (since it will be dismissed as default behavior by the popover controller) we need to manually
dismiss it if we are trying to programmatically dismiss the action sheet/menu.

* UIProcess/ios/WKActionSheet.mm:
(-[WKActionSheet presentSheetFromRect:]):
(-[WKActionSheet doneWithSheet]):
* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView setupInteraction]):
(-[WKContentView resignFirstResponder]):


Identifier: 184846@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211643 268f45cc-cd09-0410-ab3c-d52691b4dbfc