The VMInspector should use an RAII Locker. https://bugs.webkit.org/show_bug.cgi?id=167854
Reviewed by Saam Barati.

Previously, VMInspector::lock() was returning an expected LockToken, and there's
no way to unlock it when we're done with it.  This was not a problem before
because the VMInspector had only one client, the SigillCrashAnalyzer, that
expected the process to crash due to a SIGILL shortly thereafter.

However, the VMInspector is useful as a debugging tool that we can apply in other
debugging tasks.  Fixing VMInspector::lock() to return an RAII locker will enable
other use cases.  Plus it's just bad form to be able to lock something and never
be able to unlock it.

* tools/SigillCrashAnalyzer.cpp:
(JSC::SigillCrashAnalyzer::analyze):
* tools/VMInspector.cpp:
* tools/VMInspector.h:


Identifier: 184887@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211684 268f45cc-cd09-0410-ab3c-d52691b4dbfc