[GTK] Test fast/forms/select-empty-option-height.html fails https://bugs.webkit.org/show_bug.cgi?id=167985
Reviewed by Michael Catanzaro.

This seems to be a problem with the differences in scroll step or the size of the items in the list. The thing
is that 10 attempts with an scroll amount of 10 is not enough in the GTk+ port to scroll the list to the desired
point. Either increasing the scroll amount or the maximum attempts would fix the test. Increasing the scroll
amount would break the test in other ports so better use a higher number of maximum attempts.

* fast/forms/listbox-respects-padding-bottom.html:

Identifier: 185012@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211872 268f45cc-cd09-0410-ab3c-d52691b4dbfc