Don't migrate WebKit DOM headers in MigrateHeaders.make https://bugs.webkit.org/show_bug.cgi?id=167608
Source/WebKit:

Reviewed by Darin Adler.

Stop migrating DOM headers that are now in WebKit.

* WebKit.xcodeproj/project.pbxproj: Marked DOM headers as private, so that Xcode
takes care of making them available as SPI.

Source/WebKit/mac:

Reviewed by Darin Adler.

Stop migrating DOM headers that are now in WebKit.

* DOM/DOMDOMImplementation.h: Removed.
* DOM/DOMDOMImplementation.mm: Removed.
* DOM/DOMDOMImplementationInternal.h: Removed.
* DOM/DOMDOMTokenList.h: Removed.
* DOM/DOMDOMTokenList.mm: Removed.
* DOM/DOMDOMTokenListInternal.h: Removed.
* DOM/DOMImplementation.h: Copied from Source/WebKit/mac/DOM/DOMDOMImplementation.h.
* DOM/DOMImplementation.mm: Copied from Source/WebKit/mac/DOM/DOMDOMImplementation.mm.
* DOM/DOMImplementationInternal.h: Copied from Source/WebKit/mac/DOM/DOMDOMImplementationInternal.h.
* DOM/DOMTokenList.h: Copied from Source/WebKit/mac/DOM/DOMDOMTokenList.h.
* DOM/DOMTokenList.mm: Copied from Source/WebKit/mac/DOM/DOMDOMTokenList.mm.
* DOM/DOMTokenListInternal.h: Copied from Source/WebKit/mac/DOM/DOMDOMTokenListInternal.h.
As the files are not in WebCore any more, there is no need to differentiate from
WebCore internal headers with the same name. This makes it possible to make these
headers private using Xcode mechanisms, as opposed to copying with a shell script.

* DOM/DOMElement.mm: Removed "using namespace WebCore". Other DOM implementation
files don't have it.

* DOM/DOMCore.h:
* DOM/DOMDocument.mm:
* DOM/DOMHTMLAnchorElement.mm:
* DOM/DOMHTMLAreaElement.mm:
* DOM/DOMHTMLLinkElement.mm:
* DOM/DOMUtility.mm:
No more "DOMDOM".

* WebKit.exp:
No more "DOMDOM". DOMDOMTokenList was not directly exposed via an SPI header, so
renaming it is likely to be safe.

* MigrateHeaders.make: Don't migrate WebKit headers.

* Configurations/WebKitLegacy.xcconfig: Don't include two iOS specific SPI headers
on macOS.



Identifier: 185060@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211931 268f45cc-cd09-0410-ab3c-d52691b4dbfc