[GTK] scroll with transparent background not repainted after scrollY >= 32768 https://bugs.webkit.org/show_bug.cgi?id=154283
Reviewed by Carlos Garcia Campos.

Due to a limitation of the pixman backend, which uses 16 bits to hold signed integers, cairo is
not able to draw anything when using transformation matrices with values bigger than 32768. When
drawing patterns into large pages, the matrices values can overflow those 16 bits, so cairo doesn't
draw anything in, which causes the reported transparent backgrounds.

The patch modifies the transformation matrices both from the current context and the pattern we
are painting, to avoid them to hold values that cannot stored in 16 bits.

There's still the possibility that this happens, but it would require using a pattern with a size
bigger than 32768.

Based on a previous patch by Gwang Yoon Hwang  <yoon@igalia.com>.

No new tests.

* platform/graphics/cairo/CairoUtilities.cpp:
(WebCore::drawPatternToCairoContext):


Identifier: 185083@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211967 268f45cc-cd09-0410-ab3c-d52691b4dbfc