worker.postMessage should throw a TypeError if a SharedArrayBuffer is in the transfer list https://bugs.webkit.org/show_bug.cgi?id=168277
Reviewed by Mark Lam.
Source/WebCore:


Test: workers/sab/postMessage-transfer-type-error.html

This is a simple spec compliance change. The title says it all.

* bindings/js/SerializedScriptValue.cpp:
(WebCore::SerializedScriptValue::create):

LayoutTests:


Add a test of the new behavior, and remove tests for the old behavior.

Most of the SharedArrayBuffer tests use the new style, where the buffer is not in the
transfer list, and the tests being removed are clones of the no-transfer tests. So, we
aren't losing any coverage.

* workers/sab/null-worker.js: Added.
* workers/sab/postMessage-transfer-type-error-expected.txt: Added.
* workers/sab/postMessage-transfer-type-error.html: Added.
* workers/sab/sab-creator-transfer.js: Removed. (Sibling: sab-creator-no-transfer.js)
* workers/sab/sent-from-worker-transfer.html: Removed. (Sibling: sent-from-worker-no-transfer.html)
* workers/sab/simple.html: Removed. (Sibling: no-transfer.html)



Identifier: 185335@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212301 268f45cc-cd09-0410-ab3c-d52691b4dbfc