[EME] MediaKeys::setServerCertificate() must resolve with 'false' when certificates aren't supported https://bugs.webkit.org/show_bug.cgi?id=168362
Reviewed by Xabier Rodriguez-Calvar.

Source/WebCore:

In MediaKeys::setServerCertificate(), when the CDM implementation
doesn't support server certificates, it should resolve the promise
with the 'false' value, as mandated by the specification, and not
reject it.

A test case in mock-MediaKeys-setServerCertificate.html is updated.

* Modules/encryptedmedia/MediaKeys.cpp:
(WebCore::MediaKeys::setServerCertificate):

LayoutTests:

Update the certificates-not-supported test case, making sure the promise
is resolved with the 'false' result.

* media/encrypted-media/mock-MediaKeys-setServerCertificate-expected.txt:
* media/encrypted-media/mock-MediaKeys-setServerCertificate.html:


Identifier: 185383@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212356 268f45cc-cd09-0410-ab3c-d52691b4dbfc